added xournalpp
The manifest has been taken from the upstream repository.
It has been tweaked a little bit: The parallel build of libportaudiocpp has been disabled, because it failed for me.
And the empty build-options (and configure) array has been removed as well as an appdata file is installed.
The appdata file has been taken from https://github.com/xournalpp/xournalpp/issues/1298 and will be proposed as a PR upstream.